FLUXDOCTOR

+++ run.sh
@@ -1,49 +1,73 @@
 #!/bin/bash
-
+#
+# FLUXDOCTOR v1.0
+# 
 set -eu
 
-# dasm
-#   https://dasm-assembler.github.io/
+if ! type -p dasm >/dev/null 2>&1; then
+    echo "ERROR, missing executable:" 1>&2
+    echo "  dasm" 1>&2
+    echo "" 1>&2
+    echo "Please install dasm from:" 1>&2
+    echo "  https://dasm-assembler.github.io/" 1>&2
+    exit 1
+fi
 
-# Apple Commander
-#   https://applecommander.github.io/ac/
+if ! type -p AppleWin >/dev/null 2>&1; then
+    echo "ERROR, missing executable:" 1>&2
+    echo "  AppleWin" 1>&2
+    echo "" 1>&2
+    echo "Please install AppleWin from:" 1>&2
+    echo "  https://github.com/AppleWin/AppleWin" 1>&2
+    exit 1
+fi
+
 AC="$HOME/ac-windows-amd64-13.1.exe"
-
-# AppleWin
-#   https://github.com/AppleWin/AppleWin
+if [[ ! -x "$AC" ]]; then
+    echo "ERROR missing executable:" 1>&2
+    echo "  $AC" 1>&2
+    echo "" 1>&2
+    echo "Please install Apple Commander from:" 1>&2
+    echo "  https://applecommander.github.io/ac/" 1>&2
+    exit 1
+fi
 
 echo "Creating blank image:"
-echo "  images/hello.do -> diag.do"
-rm -f diag.do
-# $AC -dos140 diag.do
-cp images/hello.do diag.do
-# ls -l diag.do
+echo "  images/hello.do -> fluxdoctor.do"
+rm -f fluxdoctor.do
+# $AC -dos140 fluxdoctor.do
+cp images/hello.do fluxdoctor.do
+# ls -l fluxdoctor.do
 
 echo "Compiling:"
-echo "  apple-disk-diag.dasm -> diag.bin"
-rm -f diag.bin
-dasm apple-disk-diag.dasm -f3 -odiag.bin
-# ls -l diag.bin
+echo "  fluxdoctor.dasm -> fluxdoctor.bin"
+if [[ ! -r fluxdoctor.dasm ]]; then
+    echo "MISSING: fluxdoctor.dasm" 1>&2
+    exit 1
+fi
+rm -f fluxdoctor.bin
+dasm fluxdoctor.dasm -f3 -ofluxdoctor.bin
+# ls -l fluxdoctor.bin
 
 rm -f pgm.bin
 echo "Extracting binary (removing 4 byte header):"
-echo "  diag.bin -> pgm.bin"
+echo "  fluxdoctor.bin -> pgm.bin"
 # Get start address in hex from first two byte.
-startadr="0x$(od -An -t x2 -N2 diag.bin | tr -d ' ')"
+startadr="0x$(od -An -t x2 -N2 fluxdoctor.bin | tr -d ' ')"
 echo "Start address: $startadr"
 # Remove first four bytes (len + start)
-cat diag.bin | tail -c+5 > pgm.bin
+cat fluxdoctor.bin | tail -c+5 > pgm.bin
 
 echo "Removing existing program:"
-echo "  DIAG"
-$AC -d diag.do DIAG
+echo "  FLUXDOCTOR"
+$AC -d fluxdoctor.do FLUXDOCTOR
 
 echo "Adding binary to disk image:"
-echo "  pgm.bin -> DIAG"
-$AC -p diag.do DIAG B $startadr < pgm.bin
+echo "  pgm.bin -> FLUXDOCTOR"
+$AC -p fluxdoctor.do FLUXDOCTOR B $startadr < pgm.bin
 
 echo "Final disk image:"
-echo "  diag.do"
-$AC -ll diag.do
+echo "  fluxdoctor.do"
+$AC -ll fluxdoctor.do
 
-AppleWin -d1 diag.do
+AppleWin -d1 fluxdoctor.do
